refactor collectors into subpackage
