+++ connector/collector/docker.go
@@ -1,16 +1,17 @@
-package metrics
+package collector
 
 import (
+	"github.com/bcicen/ctop/metrics"
 	api "github.com/fsouza/go-dockerclient"
 )
 
 // Docker collector
 type Docker struct {
-	Metrics
+	metrics.Metrics
 	id         string
 	client     *api.Client
 	running    bool
-	stream     chan Metrics
+	stream     chan metrics.Metrics
 	done       chan bool
 	lastCpu    float64
 	lastSysCpu float64
@@ -18,7 +19,7 @@ type Docker struct {
 
 func NewDocker(client *api.Client, id string) *Docker {
 	return &Docker{
-		Metrics: Metrics{},
+		Metrics: metrics.Metrics{},
 		id:      id,
 		client:  client,
 	}
@@ -26,7 +27,7 @@ func NewDocker(client *api.Client, id string) *Docker {
 
 func (c *Docker) Start() {
 	c.done = make(chan bool)
-	c.stream = make(chan Metrics)
+	c.stream = make(chan metrics.Metrics)
 	stats := make(chan *api.Stats)
 
 	go func() {
@@ -60,7 +61,7 @@ func (c *Docker) Running() bool {
 	return c.running
 }
 
-func (c *Docker) Stream() chan Metrics {
+func (c *Docker) Stream() chan metrics.Metrics {
 	return c.stream
 }
