+++ metrics/main.go
@@ -1,13 +1,5 @@
 package metrics
 
-import (
-	"math"
-
-	"github.com/bcicen/ctop/logging"
-)
-
-var log = logging.Init()
-
 type Metrics struct {
 	CPUUtil      int
 	NetTx        int64
@@ -38,16 +30,4 @@ type Collector interface {
 	Running() bool
 	Start()
 	Stop()
-}
-
-func round(num float64) int {
-	return int(num + math.Copysign(0.5, num))
-}
-
-// return rounded percentage
-func percent(val float64, total float64) int {
-	if total <= 0 {
-		return 0
-	}
-	return round((val / total) * 100)
 }
